Packagist Ready
In this commit the repos should be ready for Packagist.

I have moved the Code into the src folder and I have made the class aparat of the MP3File namespace and the class remains MP3File.

The `composer.json` file has been updated with the PSR4 autoload method and testing was done to verify it does work.

`vendor\*` was for testing.
`vendor\test.php` has a test setup to check for the MP3File class is loaded.

I've added myself as an author since I have been doing updates to the repos. It can be taken out in the end.